Refactor: MVC 패턴에 맞는 의존관계를 생각해 Model 쪽으로 DTO 패키지를 분리한다.
