+++ src/main/java/org/prgms/springbootbasic/controller/HomeController.java
@@ -1,7 +1,7 @@
 package org.prgms.springbootbasic.controller;
 
 import lombok.extern.slf4j.Slf4j;
-import org.prgms.springbootbasic.controller.voucher.dto.VoucherResponseDto;
+import org.prgms.springbootbasic.service.dto.VoucherResponseDto;
 import org.prgms.springbootbasic.domain.customer.Customer;
 import org.prgms.springbootbasic.service.CustomerService;
 import org.prgms.springbootbasic.service.VoucherService;
