+++ src/main/java/org/prgms/springbootbasic/controller/voucher/VoucherController.java
@@ -1,9 +1,9 @@
 package org.prgms.springbootbasic.controller.voucher;
 
 import lombok.extern.slf4j.Slf4j;
-import org.prgms.springbootbasic.controller.voucher.dto.VoucherResponseDto;
+import org.prgms.springbootbasic.service.dto.VoucherResponseDto;
 import org.prgms.springbootbasic.domain.VoucherType;
-import org.prgms.springbootbasic.controller.voucher.dto.VoucherRequestDto;
+import org.prgms.springbootbasic.service.dto.VoucherRequestDto;
 import org.prgms.springbootbasic.exception.EntityNotFoundException;
 import org.prgms.springbootbasic.service.VoucherService;
 import org.springframework.stereotype.Controller;
