+++ src/main/java/org/prgms/springbootbasic/service/VoucherService.java
@@ -2,7 +2,7 @@ package org.prgms.springbootbasic.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.prgms.springbootbasic.service.dto.VoucherFilterDto;
-import org.prgms.springbootbasic.controller.voucher.dto.VoucherResponseDto;
+import org.prgms.springbootbasic.service.dto.VoucherResponseDto;
 import org.prgms.springbootbasic.domain.VoucherType;
 import org.prgms.springbootbasic.domain.voucher.Voucher;
 import org.prgms.springbootbasic.domain.voucher.VoucherPolicy;
