+++ src/main/java/org/prgms/springbootbasic/service/dto/VoucherRequestDto.java
@@ -1,4 +1,4 @@
-package org.prgms.springbootbasic.controller.voucher.dto;
+package org.prgms.springbootbasic.service.dto;
 
 import java.util.UUID;
