+++ src/test/java/org/prgms/springbootbasic/controller/voucher/VoucherApiControllerTest.java
@@ -3,8 +3,8 @@ package org.prgms.springbootbasic.controller.voucher;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.prgms.springbootbasic.controller.voucher.dto.VoucherRequestDto;
-import org.prgms.springbootbasic.controller.voucher.dto.VoucherResponseDto;
+import org.prgms.springbootbasic.service.dto.VoucherRequestDto;
+import org.prgms.springbootbasic.service.dto.VoucherResponseDto;
 import org.prgms.springbootbasic.domain.VoucherType;
 import org.prgms.springbootbasic.service.VoucherService;
 import org.prgms.springbootbasic.service.dto.VoucherFilterDto;
